+++ backend/foodgram/core/serializers.py
@@ -20,7 +20,7 @@ class RecipeIngredientSerializer(serializers.ModelSerializer):
     id = serializers.IntegerField(source='ingredient.id')
     name = serializers.CharField(source='ingredient.name')
     measurement_unit = serializers.CharField(source='ingredient.measurement_unit')
-    amount = serializers.FloatField()
+    amount = serializers.IntegerField()
 
     class Meta:
         model = RecipeIngredient
